Create password.csv
Here, in this program i have created a password.csv file using which i have trained my model. 
so, in that file i have used password up to 30 characters long and i used 1,2,3,4,5 to denote the strength of the password.

But if you want use password above 30 characters in length then it will be try to predict is strength but in order to increase its accuracy you can added more data to CSV file.